requirements.txt
# Reddit → Neo4j Graph setup dependencies
praw==7.7.1
update-checker==0.18.0
requests==2.31.0
python-dotenv==1.0.1
neo4j==5.25.0
tqdm==4.66.4
